Move GitFlowCheckConfiguration into GitFlow type
